dev: adjust file structure
